Add files via upload
In this project, I worked with Python functions and file handling. The main objective was to apply fundamental concepts of structured programming, creating reusable functions and managing the reading and writing of text files. This exercise reinforces programming logic and the modular organization of code, key aspects for developing more complex projects.